streamlit/requirements-legal.txt
boto3==1.26.73
pandas==1.5.3
Pillow==9.4.0
streamlit==1.20.0
torch==1.13.1
torchvision==0.14.1
PyYAML
langchain==0.0.249 
pypdf>=3.8,<4
opensearch-py
dependencies/awscli-1.29.21-py3-none-any.whl
dependencies/boto3-1.28.21-py3-none-any.whl
dependencies/botocore-1.31.21-py3-none-any.whl
